Move award_event_store.py to repositories/ & update imports
